Rolled Metadata/MetadataItem structs directly into deepspeech.h
